refactor(ui): split terminal tauri services
