+++ src/ui/src/services/tauri/pty.ts
@@ -1,1 +1,60 @@
-export {};
+import { invoke } from "@tauri-apps/api/core";
+
+export function spawnPty(
+  workingDir: string,
+  workspaceName: string,
+  workspaceId: string,
+  rootPath: string,
+  defaultBranch: string,
+  branchName: string,
+): Promise<number> {
+  return invoke("spawn_pty", {
+    workingDir,
+    workspaceName,
+    workspaceId,
+    rootPath,
+    defaultBranch,
+    branchName,
+  });
+}
+
+export function writePty(ptyId: number, data: number[]): Promise<void> {
+  return invoke("write_pty", { ptyId, data });
+}
+
+export function resizePty(
+  ptyId: number,
+  cols: number,
+  rows: number
+): Promise<void> {
+  return invoke("resize_pty", { ptyId, cols, rows });
+}
+
+export function closePty(ptyId: number): Promise<void> {
+  return invoke("close_pty", { ptyId });
+}
+
+export function interruptPtyForeground(ptyId: number): Promise<void> {
+  return invoke("interrupt_pty_foreground", { ptyId });
+}
+
+export function startAgentTaskTail(
+  tabId: number,
+  outputPath: string,
+): Promise<void> {
+  return invoke("start_agent_task_tail", { tabId, outputPath });
+}
+
+export function stopAgentTaskTail(tabId: number): Promise<void> {
+  return invoke("stop_agent_task_tail", { tabId });
+}
+
+export function stopAgentBackgroundTask(
+  chatSessionId: string,
+  taskId: string,
+): Promise<void> {
+  return invoke("stop_agent_background_task", {
+    chatSessionId,
+    taskId,
+  });
+}
